kosik pre prihlasenie a session + zaklad objednavky

+++ obchod_appka_backend/app/Http/Controllers/KosikController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Product;
-use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 
@@ -14,12 +14,39 @@ class KosikController extends Controller
      */
     public function index()
     {
-        //
-        $user = User::find(2);
-
-        $products = $user->products()->with('image')->get();
-
         $cena = 0;
+        $products = [];
+
+        if ($user = Auth::user()) {
+            $products = $user->products()->with('image')->get();
+        } else {
+            $kosik = session()->get('kosik', []);
+
+            $ids = array_column($kosik, 'product_id');
+
+            $products_list = Product::whereIn('id', $ids)->with('image')->get()->keyBy('id');
+
+            $counter = 0;
+
+            $products = array_map(function ($kosik) use ($products_list, &$counter) {
+                $product = $products_list[$kosik['product_id']];
+
+                $product->pivot = (object)[
+                    'velkost' => $kosik['velkost'],
+                    'pocet' => $kosik['pocet'],
+                    'id' => $counter++
+                ];
+
+                return $product;
+            }, $kosik);
+
+            // foreach ($kosik as &$test) {
+            //     $test['pivot'] = ['velkost' => $test['velkost']];
+            // }
+        }
+
+        // return $products;
+
 
         foreach ($products as $product) {
             $cena += $product->cena * $product->pivot->pocet;
@@ -40,10 +67,25 @@ class KosikController extends Controller
     public function store(Request $request): RedirectResponse
     {
         //
-        $user = User::find(2);
-        $product = Product::find(1);
 
-        $user->products()->attach($product->id, ['pocet' => 1, 'velkost' => 40]);
+        $product_id = 2;
+        $pocet = 1;
+        $velkost = 40;
+
+
+        if ($user = Auth::user()) {
+            $product = Product::find(1);
+            $user->products()->attach($product->id, ['pocet' => $pocet, 'velkost' => $velkost]);
+        } else {
+            $key = $product_id . '_' . $velkost;
+            $kosik = session('kosik', []);
+            $kosik[$key] = [
+                'product_id' => $product_id,
+                'pocet' => $pocet,
+                'velkost' => $velkost
+            ];
+            session(['kosik' => $kosik]);
+        }
 
         return redirect('/kosik');
     }
@@ -54,21 +96,43 @@ class KosikController extends Controller
     public function update(Request $request, string $id): RedirectResponse
     {
         //
-        $user = User::find(2);
+        if ($user = Auth::user()) {
 
-        $product = $user->products()->wherePivot('id', $id);
-        $pocet = $product->first()->pivot->pocet;
+            $product = $user->products()->wherePivot('id', $id);
+            $pocet = $product->first()->pivot->pocet;
 
-        if ($request->action === 'inc') {
-            $pocet++;
-        } elseif ($request->action === 'dec') {
-            $pocet--;
-        }
+            if ($request->action === 'inc') {
+                $pocet++;
+            } elseif ($request->action === 'dec') {
+                $pocet--;
+            }
 
-        if ($pocet < 1) {
-            $product->detach();
+            if ($pocet < 1) {
+                $product->detach();
+            } else {
+                $product->first()->pivot->update(['pocet' => $pocet]);
+            }
         } else {
-            $product->first()->pivot->update(['pocet' => $pocet]);
+            $kosik = session()->get('kosik', []);
+
+            $key = array_keys($kosik)[$id];
+
+            $pocet = $kosik[$key]['pocet'];
+
+            if ($request->action === 'inc') {
+                $pocet++;
+            } elseif ($request->action === 'dec') {
+                $pocet--;
+            }
+
+
+            if ($pocet < 1) {
+                array_splice($kosik, $id, 1);
+            } else {
+                $kosik[$key]['pocet'] = $pocet;
+            }
+
+            session(['kosik' => $kosik]);
         }
 
         return redirect('/kosik');
@@ -80,8 +144,13 @@ class KosikController extends Controller
     public function destroy(string $id): RedirectResponse
     {
         //
-        $user = User::find(2);
-        $user->products()->wherePivot('id', $id)->detach();
+        if ($user = Auth::user()) {
+            $user->products()->wherePivot('id', $id)->detach();
+        } else {
+            $kosik = session()->get('kosik', []);
+            array_splice($kosik, $id, 1);
+            session(['kosik' => $kosik]);
+        }
 
         return redirect('/kosik');
     }
